pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "google-generativeai>=0.8.5",
    "gunicorn>=23.0.0",
    "pillow>=11.2.1",
    "psycopg2-binary>=2.9.10",
    "requests>=2.32.3",
    "werkzeug>=3.1.3",
]
